+++ Pods/Target Support Files/Nimble-OSX10.11/Nimble-OSX10.11-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_Nimble_OSX10_11 : NSObject
+@end
+@implementation PodsDummy_Nimble_OSX10_11
+@end