+++ SourceryRuntime/Sources/Coding.generated.swift
@@ -3,6 +3,7 @@
 
 import Foundation
 
+
 extension NSCoder {
 
     @nonobjc func decode(forKey: String) -> String? {
@@ -45,9 +46,11 @@ extension AssociatedValue: NSCoding {}
 
 extension Attribute: NSCoding {}
 
+
 extension ClosureType: NSCoding {}
 
 extension DictionaryType: NSCoding {}
+
 
 extension EnumCase: NSCoding {}
 
@@ -56,6 +59,8 @@ extension FileParserResult: NSCoding {}
 extension Method: NSCoding {}
 
 extension MethodParameter: NSCoding {}
+
+
 
 extension TemplateContext: NSCoding {}
 
@@ -72,3 +77,4 @@ extension Typealias: NSCoding {}
 extension Types: NSCoding {}
 
 extension Variable: NSCoding {}
+