+++ Templates/Tests/Context/AutoLenses.swift
@@ -1,0 +1,26 @@
+//
+//  AutoLenses.swift
+//  Templates
+//
+//  Created by Anton Domashnev on 16.05.17.
+//  Copyright © 2017 Pixle. All rights reserved.
+//
+
+import Foundation
+
+protocol AutoLenses {}
+
+struct House: AutoLenses {
+    let rooms: Room
+    let address: String
+    let size: Int
+}
+
+struct Room: AutoLenses {
+    let people: [Person]
+    let name: String
+}
+
+struct Person: AutoLenses {
+    let name: String
+}