+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - Improved `AutoCases.stencil` to use `let` instead of computed `var`
 - Updated StencilSwiftKit to 1.0.2 which includes Stencil 0.9.0
 - Adding docset to release archive
+- Add tests for bundled stencil templates
 - Moved to CocoaPods 1.2.1
 - Made Array.parallelMap's block non-escaping